refactor: remove status-less entity status usage
