+++ src/components/EntityName/EntityName.scss
